small bit field stuff
