+++ poseidon/Cargo.toml
@@ -13,7 +13,7 @@ bit-vec = "0.6.3"
 lazy_static = "1.4.0"
 
 # netptune 
-neptune = "13.0.0"
+neptune = {version = "13.0.0", features = ["abomonation"]}
 blstrs = "0.7.1"
 typenum = "1.17.0"
