+++ poseidon/src/lib.rs
@@ -3,9 +3,9 @@
 use num_bigint::BigUint;
 use num_traits::Zero;
 
-pub mod python_transcription;
 pub mod home_baked_crypto;
 pub mod implemenations;
+pub mod python_transcription;
 extern crate ff;
 
 // const FIELD_SIZE: u32 = 8;
